job probes v2
IntelliJ uses JobImpl to create a hierarchy of nested coroutine jobs for its services, projects, etc.
JobImpl was not covered by job probes, and unfortunately it can't be easily covered, since parent context is not available in its constructor.
Instead, it would be much safer to add probes into JobSupport, though we'll have to workaround the job details collection somehow. Like casting the job to a coroutine scope in a completion handler and collecting the info there. Not ideal, but it would certainly cover all the case